requirements.txt
numpy==1.6.2
biopython==1.61
django==1.5.1

